Struct srml_support::dispatch::fmt::DebugMap

#[must_use = "must eventually call `finish()` on Debug builders"]
pub struct DebugMap<'a, 'b> where
    'b: 'a,
    'b: 'a,
    'b: 'b,  { /* fields omitted */ }

A struct to help with fmt::Debug implementations.

This is useful when you wish to output a formatted map as a part of your Debug::fmt implementation.

This can be constructed by the Formatter::debug_map method.

Examples

use std::fmt;

struct Foo(Vec<(String, i32)>);

impl fmt::Debug for Foo {
    fn fmt(&self, fmt: &mut fmt::Formatter) -> fmt::Result {
        fmt.debug_map().entries(self.0.iter().map(|&(ref k, ref v)| (k, v))).finish()
    }
}

assert_eq!(
    format!("{:?}", Foo(vec![("A".to_string(), 10), ("B".to_string(), 11)])),
    "{\"A\": 10, \"B\": 11}",
);

Methods

impl<'a, 'b> DebugMap<'a, 'b> where
'b: 'a,

pub fn entry(
&mut self,
key: &dyn Debug,
value: &dyn Debug
) -> &mut DebugMap<'a, 'b>

Adds a new entry to the map output.

Examples

use std::fmt;

struct Foo(Vec<(String, i32)>);

impl fmt::Debug for Foo {
    fn fmt(&self, fmt: &mut fmt::Formatter<'_>) -> fmt::Result {
        fmt.debug_map()
           .entry(&"whole", &self.0) // We add the "whole" entry.
           .finish()
    }
}

assert_eq!(
    format!("{:?}", Foo(vec![("A".to_string(), 10), ("B".to_string(), 11)])),
    "{\"whole\": [(\"A\", 10), (\"B\", 11)]}",
);

pub fn entries<K, V, I>(&mut self, entries: I) -> &mut DebugMap<'a, 'b> where
I: IntoIterator<Item = (K, V)>,
K: Debug,
V: Debug,

Adds the contents of an iterator of entries to the map output.

Examples

use std::fmt;

struct Foo(Vec<(String, i32)>);

impl fmt::Debug for Foo {
    fn fmt(&self, fmt: &mut fmt::Formatter<'_>) -> fmt::Result {
        fmt.debug_map()
           // We map our vec so each entries' first field will become
           // the "key".
           .entries(self.0.iter().map(|&(ref k, ref v)| (k, v)))
           .finish()
    }
}

assert_eq!(
    format!("{:?}", Foo(vec![("A".to_string(), 10), ("B".to_string(), 11)])),
    "{\"A\": 10, \"B\": 11}",
);

pub fn finish(&mut self) -> Result<(), Error>

Finishes output and returns any error encountered.

Examples

use std::fmt;

struct Foo(Vec<(String, i32)>);

impl fmt::Debug for Foo {
    fn fmt(&self, fmt: &mut fmt::Formatter<'_>) -> fmt::Result {
        fmt.debug_map()
           .entries(self.0.iter().map(|&(ref k, ref v)| (k, v)))
           .finish() // Ends the struct formatting.
    }
}

assert_eq!(
    format!("{:?}", Foo(vec![("A".to_string(), 10), ("B".to_string(), 11)])),
    "{\"A\": 10, \"B\": 11}",
);