Module srml_support::metadata

Structs

DefaultByteGetter
Wrapper over dyn pointer for accessing a cached once byte value.

FnEncode
Newtype wrapper for support encoding functions (actual the result of the function).

ModuleMetadata
All metadata about an runtime module.

RuntimeMetadataPrefixed
Metadata prefixed by a u32 for reserved usage

RuntimeMetadataV6
The metadata of a runtime.

StorageEntryMetadata
All the metadata about one storage entry.

Enums

DecodeDifferent
A type that decodes to a different type than it encodes. The user needs to make sure that both types use the same encoding.

RuntimeMetadata
The metadata of a runtime. The version ID encoded/decoded through the enum nature of RuntimeMetadata.

StorageEntryModifier
A storage entry modifier.

StorageEntryType
A storage entry type.

StorageHasher
Hasher used by storage maps

Traits

DefaultByte
A technical trait to store lazy initiated vec value as static dyn pointer.