Module srml_support::storage

Stuff to do with the runtime's storage.

Modules

child
child storage NOTE could replace unhashed by having only one kind of storage (root being null storage key (storage_key can become Option<&u8>). This module is a currently only a variant of unhashed with additional storage_key. Note that storage_key must be unique and strong (strong in the sense of being long enough to avoid collision from a resistant hash function (which unique implies)).

hashed
Operation on runtime storage using hashed keys.

storage_items
Strongly typed wrappers around values in storage.

unhashed
Operation on unhashed runtime storage

Structs

RuntimeStorage
The underlying runtime storage.

Traits

AppendableStorageMap
A storage map with values that can be appended to.

EnumerableStorageMap
A storage map that can be enumerated.

StorageDoubleMap
An implementation of a map with a two keys.

StorageMap
A strongly-typed map in storage.

StorageValue
A trait for working with macro-generated storage values under the substrate storage API.