Module srml_support::storage::child

child storage NOTE could replace unhashed by having only one kind of storage (root being null storage key (storage_key can become Option<&u8>). This module is a currently only a variant of unhashed with additional storage_key. Note that storage_key must be unique and strong (strong in the sense of being long enough to avoid collision from a resistant hash function (which unique implies)).

Re-exports

pub use super::unhashed::StorageVec;

Functions

exists
Check to see if key has an explicit entry in storage.

get
Return the value of the item in storage under key, or None if there is no explicit entry.

get_or
Return the value of the item in storage under key, or default_value if there is no explicit entry.

get_or_default
Return the value of the item in storage under key, or the type's default if there is no explicit entry.

get_or_else
Return the value of the item in storage under key, or default_value() if there is no explicit entry.

get_raw
Get a Vec of bytes from storage.

kill
Ensure key has no explicit entry in storage.

kill_storage
Remove all storage_key key/values

put
Put value in storage under key.

put_raw
Put a raw byte slice into storage.

take
Remove key from storage, returning its value if it had an explicit entry or None otherwise.

take_or
Return the value of the item in storage under key, or default_value if there is no explicit entry. Ensure there is no explicit entry on return.

take_or_default
Remove key from storage, returning its value, or, if there was no explicit entry in storage, the default for its type.

take_or_else
Return the value of the item in storage under key, or default_value() if there is no explicit entry. Ensure there is no explicit entry on return.