Trait srml_support::storage::hashed::generator::StorageList

pub trait StorageList<T: Codec> {
    fn prefix() -> &'static [u8];

    fn len_key() -> Vec<u8>;

    fn key_for(index: u32) -> Vec<u8>;

    fn items<S: HashedStorage<Twox128>>(storage: &S) -> Vec<T>;

    fn set_items<S: HashedStorage<Twox128>>(items: &[T], storage: &S);

    fn set_item<S: HashedStorage<Twox128>>(index: u32, item: &T, storage: &S);

    fn get<S: HashedStorage<Twox128>>(index: u32, storage: &S) -> Option<T>;

    fn len<S: HashedStorage<Twox128>>(storage: &S) -> u32;

    fn clear<S: HashedStorage<Twox128>>(storage: &S);
}

A strongly-typed list in storage.

Required methods

Important traits for &'_ mut [u8]
impl<'_> Write for &'_ mut [u8]impl<'_> Read for &'_ [u8]

fn prefix() -> &'static [u8]

Get the prefix key in storage.

Important traits for Vec<u8>
impl Write for Vec<u8>

fn len_key() -> Vec<u8>

Get the key used to put the length field.

Important traits for Vec<u8>
impl Write for Vec<u8>

fn key_for(index: u32) -> Vec<u8>

Get the storage key used to fetch a value at a given index.

Important traits for Vec<u8>
impl Write for Vec<u8>

fn items<S: HashedStorage<Twox128>>(storage: &S) -> Vec<T>

Read out all the items.

fn set_items<S: HashedStorage<Twox128>>(items: &[T], storage: &S)

Set the current set of items.

fn set_item<S: HashedStorage<Twox128>>(index: u32, item: &T, storage: &S)

Set the item at the given index.

fn get<S: HashedStorage<Twox128>>(index: u32, storage: &S) -> Option<T>

Load the value at given index. Returns None if the index is out-of-bounds.

fn len<S: HashedStorage<Twox128>>(storage: &S) -> u32

Load the length of the list

fn clear<S: HashedStorage<Twox128>>(storage: &S)

Clear the list.

Loading content...