Trait srml_support::storage::hashed::generator::StorageValue

pub trait StorageValue<T: Codec> {
    type Query;
    fn key() -> &'static [u8];

    fn get<S: HashedStorage<Twox128>>(storage: &S) -> Self::Query;

    fn take<S: HashedStorage<Twox128>>(storage: &mut S) -> Self::Query;

    fn mutate<R, F: FnOnce(&mut Self::Query) -> R, S: HashedStorage<Twox128>>(
        f: F, 
        storage: &mut S
    ) -> R;

    fn exists<S: HashedStorage<Twox128>>(storage: &S) -> bool { ... }

    fn put<S: HashedStorage<Twox128>>(val: &T, storage: &mut S) { ... }

    fn kill<S: HashedStorage<Twox128>>(storage: &mut S) { ... }

    fn append<S: HashedStorage<Twox128>, I: Encode>(
        items: &[I], 
        storage: &mut S
    ) -> Result<(), &'static str>
    where
        T: EncodeAppend<Item = I>,
    { ... }
}

A strongly-typed value kept in storage.

Associated Types

type Query

The type that get/take returns.

Loading content...
Required methods

Important traits for &'_ mut [u8]
impl<'_> Write for &'_ mut [u8]impl<'_> Read for &'_ [u8]

fn key() -> &'static [u8]

Get the storage key.

fn get<S: HashedStorage<Twox128>>(storage: &S) -> Self::Query

Load the value from the provided storage instance.

fn take<S: HashedStorage<Twox128>>(storage: &mut S) -> Self::Query

Take a value from storage, removing it afterwards.

fn mutate<R, F: FnOnce(&mut Self::Query) -> R, S: HashedStorage<Twox128>>(
f: F,
storage: &mut S
) -> R

Mutate this value

Loading content...
Provided methods

fn exists<S: HashedStorage<Twox128>>(storage: &S) -> bool

true if the value is defined in storage.

fn put<S: HashedStorage<Twox128>>(val: &T, storage: &mut S)

Store a value under this key into the provided storage instance.

fn kill<S: HashedStorage<Twox128>>(storage: &mut S)

Clear the storage value.

fn append<S: HashedStorage<Twox128>, I: Encode>(
items: &[I],
storage: &mut S
) -> Result<(), &'static str> where
T: EncodeAppend<Item = I>,

Append the given items to the value in the storage.

T is required to implement codec::EncodeAppend.

Loading content...