Trait srml_support::storage::StorageDoubleMap

pub trait StorageDoubleMap<K1: Encode, K2: Encode, V: Codec> {
    type Query;
    fn exists<KArg1: ?Sized, KArg2: ?Sized>(k1: &KArg1, k2: &KArg2) -> bool
    where
        K1: Borrow<KArg1>,
        K2: Borrow<KArg2>,
        KArg1: Encode,
        KArg2: Encode;

    fn get<KArg1: ?Sized, KArg2: ?Sized>(k1: &KArg1, k2: &KArg2) -> Self::Query
    where
        K1: Borrow<KArg1>,
        K2: Borrow<KArg2>,
        KArg1: Encode,
        KArg2: Encode;

    fn take<KArg1: ?Sized, KArg2: ?Sized>(k1: &KArg1, k2: &KArg2) -> Self::Query
    where
        K1: Borrow<KArg1>,
        K2: Borrow<KArg2>,
        KArg1: Encode,
        KArg2: Encode;

    fn insert<KArg1: ?Sized, KArg2: ?Sized, VArg: ?Sized>(
        k1: &KArg1, 
        k2: &KArg2, 
        val: &VArg
    )
    where
        K1: Borrow<KArg1>,
        K2: Borrow<KArg2>,
        V: Borrow<VArg>,
        KArg1: Encode,
        KArg2: Encode,
        VArg: Encode;

    fn remove<KArg1: ?Sized, KArg2: ?Sized>(k1: &KArg1, k2: &KArg2)
    where
        K1: Borrow<KArg1>,
        K2: Borrow<KArg2>,
        KArg1: Encode,
        KArg2: Encode;

    fn remove_prefix<KArg1: ?Sized>(k1: &KArg1)
    where
        KArg1: Encode,
        K1: Borrow<KArg1>;

    fn mutate<KArg1: ?Sized, KArg2: ?Sized, R, F>(
        k1: &KArg1, 
        k2: &KArg2, 
        f: F
    ) -> R
    where
        K1: Borrow<KArg1>,
        K2: Borrow<KArg2>,
        KArg1: Encode,
        KArg2: Encode,
        F: FnOnce(&mut Self::Query) -> R;

    fn append<KArg1: ?Sized, KArg2: ?Sized, I>(
        k1: &KArg1, 
        k2: &KArg2, 
        items: &[I]
    ) -> Result<(), &'static str>
    where
        K1: Borrow<KArg1>,
        K2: Borrow<KArg2>,
        KArg1: Encode,
        KArg2: Encode,
        I: Encode,
        V: EncodeAppend<Item = I>;
}

An implementation of a map with a two keys.

It provides an important ability to efficiently remove all entries that have a common first key.

Associated Types

type Query

The type that get/take returns.

Loading content...
Required methods

fn exists<KArg1: ?Sized, KArg2: ?Sized>(k1: &KArg1, k2: &KArg2) -> bool where
K1: Borrow<KArg1>,
K2: Borrow<KArg2>,
KArg1: Encode,
KArg2: Encode,

fn get<KArg1: ?Sized, KArg2: ?Sized>(k1: &KArg1, k2: &KArg2) -> Self::Query where
K1: Borrow<KArg1>,
K2: Borrow<KArg2>,
KArg1: Encode,
KArg2: Encode,

fn take<KArg1: ?Sized, KArg2: ?Sized>(k1: &KArg1, k2: &KArg2) -> Self::Query where
K1: Borrow<KArg1>,
K2: Borrow<KArg2>,
KArg1: Encode,
KArg2: Encode,

fn insert<KArg1: ?Sized, KArg2: ?Sized, VArg: ?Sized>(
k1: &KArg1,
k2: &KArg2,
val: &VArg
) where
K1: Borrow<KArg1>,
K2: Borrow<KArg2>,
V: Borrow<VArg>,
KArg1: Encode,
KArg2: Encode,
VArg: Encode,

fn remove<KArg1: ?Sized, KArg2: ?Sized>(k1: &KArg1, k2: &KArg2) where
K1: Borrow<KArg1>,
K2: Borrow<KArg2>,
KArg1: Encode,
KArg2: Encode,

fn remove_prefix<KArg1: ?Sized>(k1: &KArg1) where
KArg1: Encode,
K1: Borrow<KArg1>,

fn mutate<KArg1: ?Sized, KArg2: ?Sized, R, F>(k1: &KArg1, k2: &KArg2, f: F) -> R where
K1: Borrow<KArg1>,
K2: Borrow<KArg2>,
KArg1: Encode,
KArg2: Encode,
F: FnOnce(&mut Self::Query) -> R,

fn append<KArg1: ?Sized, KArg2: ?Sized, I>(
k1: &KArg1,
k2: &KArg2,
items: &[I]
) -> Result<(), &'static str> where
K1: Borrow<KArg1>,
K2: Borrow<KArg2>,
KArg1: Encode,
KArg2: Encode,
I: Encode,
V: EncodeAppend<Item = I>,

Loading content...