Trait srml_support::storage::StorageLinkedMap

pub trait StorageLinkedMap<K: Codec, V: Codec> {
    type Query;
    type Enumerator: Iterator<Item = (K, V)>;
    fn exists<KeyArg: Borrow<K>>(key: KeyArg) -> bool;

    fn get<KeyArg: Borrow<K>>(key: KeyArg) -> Self::Query;

    fn swap<KeyArg1: Borrow<K>, KeyArg2: Borrow<K>>(
        key1: KeyArg1, 
        key2: KeyArg2
    );

    fn insert<KeyArg: Borrow<K>, ValArg: Borrow<V>>(key: KeyArg, val: ValArg);

    fn insert_ref<KeyArg: Borrow<K>, ValArg: ?Sized + Encode>(
        key: KeyArg, 
        val: &ValArg
    )
    where
        V: AsRef<ValArg>;

    fn remove<KeyArg: Borrow<K>>(key: KeyArg);

    fn mutate<KeyArg: Borrow<K>, R, F: FnOnce(&mut Self::Query) -> R>(
        key: KeyArg, 
        f: F
    ) -> R;

    fn take<KeyArg: Borrow<K>>(key: KeyArg) -> Self::Query;

    fn head() -> Option<K>;

    fn enumerate() -> Self::Enumerator;

    fn decode_len<KeyArg: Borrow<K>>(key: KeyArg) -> Result<usize, &'static str>
    where
        V: DecodeLength + Len;
}

A strongly-typed linked map in storage.

Similar to StorageMap but allows to enumerate other elements and doesn't implement append.

Associated Types

type Query

The type that get/take return.

type Enumerator: Iterator<Item = (K, V)>

The type that iterates over all (key, value).

Loading content...
Required methods

fn exists<KeyArg: Borrow<K>>(key: KeyArg) -> bool

Does the value (explicitly) exist in storage?

fn get<KeyArg: Borrow<K>>(key: KeyArg) -> Self::Query

Load the value associated with the given key from the map.

fn swap<KeyArg1: Borrow<K>, KeyArg2: Borrow<K>>(key1: KeyArg1, key2: KeyArg2)

Swap the values of two keys.

fn insert<KeyArg: Borrow<K>, ValArg: Borrow<V>>(key: KeyArg, val: ValArg)

Store a value to be associated with the given key from the map.

fn insert_ref<KeyArg: Borrow<K>, ValArg: ?Sized + Encode>(
key: KeyArg,
val: &ValArg
) where
V: AsRef<ValArg>,

Store a value under this key into the provided storage instance; this can take any reference type that derefs to T (and has Encode implemented).

fn remove<KeyArg: Borrow<K>>(key: KeyArg)

Remove the value under a key.

fn mutate<KeyArg: Borrow<K>, R, F: FnOnce(&mut Self::Query) -> R>(
key: KeyArg,
f: F
) -> R

Mutate the value under a key.

fn take<KeyArg: Borrow<K>>(key: KeyArg) -> Self::Query

Take the value under a key.

fn head() -> Option<K>

Return current head element.

fn enumerate() -> Self::Enumerator

Enumerate all elements in the map.

fn decode_len<KeyArg: Borrow<K>>(key: KeyArg) -> Result<usize, &'static str> where
V: DecodeLength + Len,

Read the length of the value in a fast way, without decoding the entire value.

T is required to implement Codec::DecodeLength.

Note that 0 is returned as the default value if no encoded value exists at the given key. Therefore, this function cannot be used as a sign of existence. use the ::exists() function for this purpose.

Loading content...