Trait srml_support::storage::unhashed::generator::StorageDoubleMap

pub trait StorageDoubleMap<K1: Codec, K2: Codec, V: Codec> {
    type Query;
    fn prefix() -> &'static [u8];

    fn key_for(k1: &K1, k2: &K2) -> Vec<u8>;

    fn prefix_for(k1: &K1) -> Vec<u8>;

    fn get<S: UnhashedStorage>(k1: &K1, k2: &K2, storage: &S) -> Self::Query;

    fn take<S: UnhashedStorage>(
        k1: &K1, 
        k2: &K2, 
        storage: &mut S
    ) -> Self::Query;

    fn mutate<R, F: FnOnce(&mut Self::Query) -> R, S: UnhashedStorage>(
        k1: &K1, 
        k2: &K2, 
        f: F, 
        storage: &mut S
    ) -> R;

    fn exists<S: UnhashedStorage>(k1: &K1, k2: &K2, storage: &S) -> bool { ... }

    fn insert<S: UnhashedStorage>(k1: &K1, k2: &K2, val: &V, storage: &mut S) { ... }

    fn remove<S: UnhashedStorage>(k1: &K1, k2: &K2, storage: &mut S) { ... }

    fn remove_prefix<S: UnhashedStorage>(k1: &K1, storage: &mut S) { ... }

    fn append<I, S: UnhashedStorage>(
        k1: &K1, 
        k2: &K2, 
        items: &[I], 
        storage: &mut S
    ) -> Result<(), &'static str>
    where
        I: Encode,
        V: EncodeAppend<Item = I>,
    { ... }
}

An implementation of a map with a two keys.

It provides an important ability to efficiently remove all entries that have a common first key.

Mapping of keys to a storage path

The storage key (i.e. the key under which the Value will be stored) is created from two parts. The first part is a hash of a concatenation of the PREFIX and Key1. And the second part is a hash of a Key2.

/!\ be careful while choosing the Hash, indeed malicious could craft second keys to lower the trie.

Associated Types

type Query

The type that get/take returns.

Loading content...
Required methods

Important traits for &'_ mut [u8]
impl<'_> Write for &'_ mut [u8]impl<'_> Read for &'_ [u8]

fn prefix() -> &'static [u8]

Get the prefix key in storage.

Important traits for Vec<u8>
impl Write for Vec<u8>

fn key_for(k1: &K1, k2: &K2) -> Vec<u8>

Get the storage key used to fetch a value corresponding to a specific key.

Important traits for Vec<u8>
impl Write for Vec<u8>

fn prefix_for(k1: &K1) -> Vec<u8>

Get the storage prefix used to fetch keys corresponding to a specific key1.

fn get<S: UnhashedStorage>(k1: &K1, k2: &K2, storage: &S) -> Self::Query

Load the value associated with the given key from the map.

fn take<S: UnhashedStorage>(k1: &K1, k2: &K2, storage: &mut S) -> Self::Query

Take the value under a key.

fn mutate<R, F: FnOnce(&mut Self::Query) -> R, S: UnhashedStorage>(
k1: &K1,
k2: &K2,
f: F,
storage: &mut S
) -> R

Mutate the value under a key.

Loading content...
Provided methods

fn exists<S: UnhashedStorage>(k1: &K1, k2: &K2, storage: &S) -> bool

true if the value is defined in storage.

fn insert<S: UnhashedStorage>(k1: &K1, k2: &K2, val: &V, storage: &mut S)

Store a value to be associated with the given key from the map.

fn remove<S: UnhashedStorage>(k1: &K1, k2: &K2, storage: &mut S)

Remove the value under a key.

fn remove_prefix<S: UnhashedStorage>(k1: &K1, storage: &mut S)

Removes all entries that shares the k1 as the first key.

fn append<I, S: UnhashedStorage>(
k1: &K1,
k2: &K2,
items: &[I],
storage: &mut S
) -> Result<(), &'static str> where
I: Encode,
V: EncodeAppend<Item = I>,

Append the given items to the value under the key specified.

Loading content...