Trait srml_support::storage::unhashed::generator::UnhashedStorage

pub trait UnhashedStorage {
    fn exists(&self, key: &[u8]) -> bool;

    fn get<T: Decode>(&self, key: &[u8]) -> Option<T>;

    fn put<T: Encode>(&mut self, key: &[u8], val: &T);

    fn kill(&mut self, key: &[u8]);

    fn kill_prefix(&mut self, prefix: &[u8]);

    fn get_raw(&self, key: &[u8]) -> Option<Vec<u8>>;

    fn put_raw(&mut self, key: &[u8], value: &[u8]);

    fn require<T: Decode>(&self, key: &[u8]) -> T { ... }

    fn get_or_default<T: Decode + Default>(&self, key: &[u8]) -> T { ... }

    fn take<T: Decode>(&mut self, key: &[u8]) -> Option<T> { ... }

    fn take_or_panic<T: Decode>(&mut self, key: &[u8]) -> T { ... }

    fn take_or_default<T: Decode + Default>(&mut self, key: &[u8]) -> T { ... }
}

Abstraction around storage with unhashed access.

Required methods

fn exists(&self, key: &[u8]) -> bool

true if the key exists in storage.

fn get<T: Decode>(&self, key: &[u8]) -> Option<T>

Load the bytes of a key from storage. Can panic if the type is incorrect.

fn put<T: Encode>(&mut self, key: &[u8], val: &T)

Put a value in under a key.

fn kill(&mut self, key: &[u8])

Remove the bytes of a key from storage.

fn kill_prefix(&mut self, prefix: &[u8])

Remove the bytes of a key from storage.

fn get_raw(&self, key: &[u8]) -> Option<Vec<u8>>

Get a Vec of bytes from storage.

fn put_raw(&mut self, key: &[u8], value: &[u8])

Put a raw byte slice into storage.

Loading content...
Provided methods

fn require<T: Decode>(&self, key: &[u8]) -> T

Load the bytes of a key from storage. Can panic if the type is incorrect. Will panic if it's not there.

fn get_or_default<T: Decode + Default>(&self, key: &[u8]) -> T

Load the bytes of a key from storage. Can panic if the type is incorrect. The type's default is returned if it's not there.

fn take<T: Decode>(&mut self, key: &[u8]) -> Option<T>

Take a value from storage, deleting it after reading.

fn take_or_panic<T: Decode>(&mut self, key: &[u8]) -> T

Take a value from storage, deleting it after reading.

fn take_or_default<T: Decode + Default>(&mut self, key: &[u8]) -> T

Take a value from storage, deleting it after reading.

Loading content...