Trait srml_support::StorageDoubleMapWithHasher

pub trait StorageDoubleMapWithHasher {
    type Key1: Codec;
    type Key2: Codec;
    type Value: Codec + Default;

    const PREFIX: &'static [u8];

    fn derive_key1(key1_data: Vec<u8>) -> Vec<u8>;

    fn derive_key2(key2_data: Vec<u8>) -> Vec<u8>;

    fn insert<Q, R>(k1: &Q, k2: &R, val: Self::Value)
    where
        Self::Key1: Borrow<Q>,
        Self::Key2: Borrow<R>,
        Q: Codec,
        R: Codec,
    { ... }

    fn remove<Q, R>(k1: &Q, k2: &R)
    where
        Self::Key1: Borrow<Q>,
        Self::Key2: Borrow<R>,
        Q: Codec,
        R: Codec,
    { ... }

    fn get<Q, R>(k1: &Q, k2: &R) -> Option<Self::Value>
    where
        Self::Key1: Borrow<Q>,
        Self::Key2: Borrow<R>,
        Q: Codec,
        R: Codec,
    { ... }

    fn exists<Q, R>(k1: &Q, k2: &R) -> bool
    where
        Self::Key1: Borrow<Q>,
        Self::Key2: Borrow<R>,
        Q: Codec,
        R: Codec,
    { ... }

    fn remove_prefix<Q>(k1: &Q)
    where
        Self::Key1: Borrow<Q>,
        Q: Codec,
    { ... }

    fn encode_key1<Q>(key: &Q) -> Vec<u8>
    where
        Self::Key1: Borrow<Q>,
        Q: Codec,
    { ... }

    fn encode_key2<R>(key: &R) -> Vec<u8>
    where
        Self::Key2: Borrow<R>,
        R: Codec,
    { ... }

    fn full_key<Q, R>(k1: &Q, k2: &R) -> Vec<u8>
    where
        Self::Key1: Borrow<Q>,
        Self::Key2: Borrow<R>,
        Q: Codec,
        R: Codec,
    { ... }
}

An implementation of a map with a two keys.

It provides an important ability to efficiently remove all entries that have a common first key.

Mapping of keys to a storage path

The storage key (i.e. the key under which the Value will be stored) is created from two parts. The first part is a hash of a concatenation of the PREFIX and Key1. And the second part is a hash of a Key2.

Hasher are implemented in derive_key* methods.

Associated Types

type Key1: Codec

type Key2: Codec

type Value: Codec + Default

Loading content...
Associated Constants

const PREFIX: &'static [u8]

Loading content...
Required methods

Important traits for Vec<u8>
impl Write for Vec<u8>

fn derive_key1(key1_data: Vec<u8>) -> Vec<u8>

Derive the first part of the key

Important traits for Vec<u8>
impl Write for Vec<u8>

fn derive_key2(key2_data: Vec<u8>) -> Vec<u8>

Derive the remaining part of the key

Loading content...
Provided methods

fn insert<Q, R>(k1: &Q, k2: &R, val: Self::Value) where
Self::Key1: Borrow<Q>,
Self::Key2: Borrow<R>,
Q: Codec,
R: Codec,

Insert an entry into this map.

fn remove<Q, R>(k1: &Q, k2: &R) where
Self::Key1: Borrow<Q>,
Self::Key2: Borrow<R>,
Q: Codec,
R: Codec,

Remove an entry from this map.

fn get<Q, R>(k1: &Q, k2: &R) -> Option<Self::Value> where
Self::Key1: Borrow<Q>,
Self::Key2: Borrow<R>,
Q: Codec,
R: Codec,

Get an entry from this map.

If there is no entry stored under the given keys, returns None.

fn exists<Q, R>(k1: &Q, k2: &R) -> bool where
Self::Key1: Borrow<Q>,
Self::Key2: Borrow<R>,
Q: Codec,
R: Codec,

Returns true if value under the specified keys exists.

fn remove_prefix<Q>(k1: &Q) where
Self::Key1: Borrow<Q>,
Q: Codec,

Removes all entries that shares the k1 as the first key.

Important traits for Vec<u8>
impl Write for Vec<u8>

fn encode_key1<Q>(key: &Q) -> Vec<u8> where
Self::Key1: Borrow<Q>,
Q: Codec,

Encode key1 into Vec and prepend a prefix

Important traits for Vec<u8>
impl Write for Vec<u8>

fn encode_key2<R>(key: &R) -> Vec<u8> where
Self::Key2: Borrow<R>,
R: Codec,

Encode key2 into Vec

Important traits for Vec<u8>
impl Write for Vec<u8>

fn full_key<Q, R>(k1: &Q, k2: &R) -> Vec<u8> where
Self::Key1: Borrow<Q>,
Self::Key2: Borrow<R>,
Q: Codec,
R: Codec,

Returns a compound key that consist of the two parts: (prefix, k1) and k2. The first part is hashed and then concatenated with a hash of k2.

Loading content...