Trait srml_support::storage::StorageDoubleMap

pub trait StorageDoubleMap<K1: Codec, K2: Codec, V: Codec> {
    type Query;
    fn prefix() -> &'static [u8];

    fn key_for<KArg1: Borrow<K1>, KArg2: Borrow<K2>>(
        k1: KArg1, 
        k2: KArg2
    ) -> Vec<u8>;

    fn prefix_for<KArg1: Borrow<K1>>(k1: KArg1) -> Vec<u8>;

    fn exists<KArg1: Borrow<K1>, KArg2: Borrow<K2>>(
        k1: KArg1, 
        k2: KArg2
    ) -> bool;

    fn get<KArg1: Borrow<K1>, KArg2: Borrow<K2>>(
        k1: KArg1, 
        k2: KArg2
    ) -> Self::Query;

    fn take<KArg1: Borrow<K1>, KArg2: Borrow<K2>>(
        k1: KArg1, 
        k2: KArg2
    ) -> Self::Query;

    fn insert<KArg1: Borrow<K1>, KArg2: Borrow<K2>, VArg: Borrow<V>>(
        k1: KArg1, 
        k2: KArg2, 
        val: VArg
    );

    fn remove<KArg1: Borrow<K1>, KArg2: Borrow<K2>>(k1: KArg1, k2: KArg2);

    fn remove_prefix<KArg1: Borrow<K1>>(k1: KArg1);

    fn mutate<KArg1, KArg2, R, F>(k1: KArg1, k2: KArg2, f: F) -> R
    where
        KArg1: Borrow<K1>,
        KArg2: Borrow<K2>,
        F: FnOnce(&mut Self::Query) -> R;

    fn append<KArg1, KArg2, I>(
        k1: KArg1, 
        k2: KArg2, 
        items: &[I]
    ) -> Result<(), &'static str>
    where
        KArg1: Borrow<K1>,
        KArg2: Borrow<K2>,
        I: Encode,
        V: EncodeAppend<Item = I>;
}

An implementation of a map with a two keys.

It provides an important ability to efficiently remove all entries that have a common first key.

Mapping of keys to a storage path

The storage key (i.e. the key under which the Value will be stored) is created from two parts. The first part is a hash of a concatenation of the PREFIX and Key1. And the second part is a hash of a Key2.

/!\ be careful while choosing the Hash, indeed malicious could craft second keys to lower the trie.

Associated Types

type Query

The type that get/take returns.

Loading content...
Required methods

Important traits for &'_ mut [u8]
impl<'_> Write for &'_ mut [u8]impl<'_> Read for &'_ [u8]

fn prefix() -> &'static [u8]

Get the prefix key in storage.

Important traits for Vec<u8>
impl Write for Vec<u8>

fn key_for<KArg1: Borrow<K1>, KArg2: Borrow<K2>>(
k1: KArg1,
k2: KArg2
) -> Vec<u8>

Get the storage key used to fetch a value corresponding to a specific key.

Important traits for Vec<u8>
impl Write for Vec<u8>

fn prefix_for<KArg1: Borrow<K1>>(k1: KArg1) -> Vec<u8>

Get the storage prefix used to fetch keys corresponding to a specific key1.

fn exists<KArg1: Borrow<K1>, KArg2: Borrow<K2>>(k1: KArg1, k2: KArg2) -> bool

true if the value is defined in storage.

fn get<KArg1: Borrow<K1>, KArg2: Borrow<K2>>(
k1: KArg1,
k2: KArg2
) -> Self::Query

Load the value associated with the given key from the map.

fn take<KArg1: Borrow<K1>, KArg2: Borrow<K2>>(
k1: KArg1,
k2: KArg2
) -> Self::Query

Take the value under a key.

fn insert<KArg1: Borrow<K1>, KArg2: Borrow<K2>, VArg: Borrow<V>>(
k1: KArg1,
k2: KArg2,
val: VArg
)

Store a value to be associated with the given key from the map.

fn remove<KArg1: Borrow<K1>, KArg2: Borrow<K2>>(k1: KArg1, k2: KArg2)

Remove the value under a key.

fn remove_prefix<KArg1: Borrow<K1>>(k1: KArg1)

Removes all entries that shares the k1 as the first key.

fn mutate<KArg1, KArg2, R, F>(k1: KArg1, k2: KArg2, f: F) -> R where
KArg1: Borrow<K1>,
KArg2: Borrow<K2>,
F: FnOnce(&mut Self::Query) -> R,

Mutate the value under a key.

fn append<KArg1, KArg2, I>(
k1: KArg1,
k2: KArg2,
items: &[I]
) -> Result<(), &'static str> where
KArg1: Borrow<K1>,
KArg2: Borrow<K2>,
I: Encode,
V: EncodeAppend<Item = I>,

Append the given items to the value under the key specified.

V is required to implement codec::EncodeAppend<Item=I>.

Loading content...